Implement MIDI thru. Reduce CPU usage in main loop
Add MIDI thru functionality. All MIDI data can be routed to a FIFO,
which can be read with a software synth to get sound out. The MIDI thru
is compile-time feature, requiring `:notewhacker-midi-thru` in
`*FEATURES*` to be compiled in.

Add documentation on how to use this MIDI thru with fluidsynth software
synthesizer.

Fix (of some kind) for 100% CPU usage. I ported this from
lispbuilder-sdl to cl-sdl2 earlier because lispbuilder-sdl was marked as
deprecated. However, lispbuilder-sdl had a built-in FPS limiter which I
used. When porting, I did a sloppy job and the main-loop was
busy-looping, checking when it was time to draw a new frame. The fix now
uses `sdl2:delay` in the idle handler to sleep about until the next
frame should be drawn. This cuts the CPU usage, but I'm not entirely
happy with it, since the delay is not exact (it'll burn some cycles,
around 1ms worth, as to not be late to start the next frame) and I'm not
sure if the sleeping causes trouble in other main handler's
branches (`:keydown`, `:keyup`, MIDI data handling etc.). In future,
maybe sdl2.kit could be used.

Some cosmetic improvements to comments and documentation.